% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapalign.R
\name{align_seq2mrf}
\alias{align_seq2mrf}
\title{align sequence to mrf}
\usage{
align_seq2mrf(
  seq,
  mrf,
  iteration = 20,
  wt_h = 1,
  wt_j = 1,
  gap_ext = 0.1,
  gap_open = -1,
  debug = TRUE
)
}
\arguments{
\item{seq}{sequence}

\item{mrf}{mrf read by read_mrf}

\item{iteration}{number of iterations}

\item{wt_h}{weight of field term H}

\item{wt_j}{weight of coupling term J}

\item{gap_ext}{gap extension penalty}

\item{gap_open}{gap open penalty}

\item{debug}{verbose}
}
\value{
a2b, mapping index to sequence. Length is the same as mrf
}
\description{
align sequence to mrf
}
