% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapalign.R
\name{bench_a2b}
\alias{bench_a2b}
\title{bencha2b}
\usage{
bench_a2b(a2b_0b, seq, mrf, seq_ref = NULL, ct_ref = NULL)
}
\arguments{
\item{a2b_0b}{a2b 0 based}

\item{seq}{sequence}

\item{mrf}{mrf from read_mrf}

\item{seq_ref}{reference sequence}

\item{ct_ref}{reference ct with BP}
}
\value{
named vector of benchmark metrics
}
\description{
bencha2b
}
