% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BP_canonical}
\alias{BP_canonical}
\title{A dataset containing canonical BP}
\format{
A character vector of canonical pairs
}
\usage{
BP_canonical
}
\description{
A dataset containing canonical BP
}
\keyword{datasets}
