% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapalign.R
\name{mrfaln_seqs}
\alias{mrfaln_seqs}
\title{mrfaln_seqs}
\usage{
mrfaln_seqs(seqs, mrf)
}
\arguments{
\item{seqs}{seqs read from seqinr::read.fasta}

\item{mrf}{mrf model read from RNAmrf::read_mrf_renum}
}
\value{
a list of re-aligned sequences in a2m format.
        Can be converted to other alignment format using esl-reformat
}
\description{
mrfaln_seqs
}
