% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{msa_a2m2seqidx}
\alias{msa_a2m2seqidx}
\title{Convert a2m formatted multiple sequence alignment to seqidx, a reference-to-sequence index}
\usage{
msa_a2m2seqidx(seq, idx_aln = NULL)
}
\arguments{
\item{seq}{a sequence in a MSA, read from seqinr::read.fasta}

\item{idx_aln}{integer vector, reference columns. if NULL, it's guessed from the uppercase alignment.}
}
\value{
seqidx, integer vector, pointer from reference columns to sequence.
e.g. seqidx[i]=5 means 5th residue is aligned to i.
0 means a gap.
}
\description{
In an a2m formatted MSA, lowercase residues are insertions, - are gaps, uppercase residues are aligned.
}
