% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapalign.R
\name{pair_a2b2aln}
\alias{pair_a2b2aln}
\title{pair_a2b2aln convert two a2b to aligned sequence.}
\usage{
pair_a2b2aln(a2b_1, a2b_2, seqs)
}
\arguments{
\item{a2b_1}{integer vector of length(seq1),encoding the alignment of seq1 to MRF}

\item{a2b_2}{integer vector of length(seq2),encoding the alignment of seq2 to MRF}

\item{seqs}{unaligned seq list, read from seqinr::read.fasta()}
}
\value{
list of aligned seqs
}
\description{
pair_a2b2aln convert two a2b to aligned sequence.
}
