% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapalign.R
\name{read_mrf}
\alias{read_mrf}
\title{read_mrf read a MRF model from GREMLIN output.}
\usage{
read_mrf(filemrf)
}
\arguments{
\item{filemrf}{}
}
\value{
list object of mrf
}
\description{
read_mrf read a MRF model from GREMLIN output.
}
