% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapalign.R
\name{read_mrf_renum}
\alias{read_mrf_renum}
\title{read_mrf_renum read a MRF model from GREMLIN output, column renumbered}
\usage{
read_mrf_renum(filemrf)
}
\arguments{
\item{filemrf}{}
}
\value{
list object of mrf
}
\description{
read_mrf_renum read a MRF model from GREMLIN output, column renumbered
}
