% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ss2pairs}
\alias{ss2pairs}
\title{ss2pairs}
\usage{
ss2pairs(ss)
}
\arguments{
\item{ss}{Character, Secondary structure in dot bracket format}
}
\value{
dataframe
           - id: 1:length(ss)
           - pair: paired id, otherwise 0
           - ss: character
}
\description{
Convert ss to a dataframe with pair info.
}
