% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sto2ss}
\alias{sto2ss}
\title{sto2ss read SS_cons from stockholm format MSA}
\usage{
sto2ss(fsto)
}
\arguments{
\item{fsto}{file path to stockholm file with GC SS_cons field}
}
\value{
ss2 Secondary structure in dot bracket format
}
\description{
If multiple lines, paste them together.
}
